+++ robot_descriptions/_descriptions.py
@@ -122,6 +122,7 @@ DESCRIPTIONS: Dict[str, Description] = {
     "skydio_x2_description": Description(Format.URDF),
     "skydio_x2_mj_description": Description(Format.MJCF),
     "so_arm100": Description(Format.URDF),
+    "so_arm100_mj_description": Description(Format.MJCF),
     "solo_description": Description(Format.URDF),
     "spryped_description": Description(Format.URDF),
     "stretch_description": Description(Format.URDF),